+++ CotEditor/Sources/Document.swift
@@ -79,7 +79,7 @@ final class Document: NSDocument, EncodingHolder {
     // MARK: Private Properties
     
     private lazy var printPanelAccessoryController: PrintPanelAccessoryController = PrintPanelAccessoryController()
-    @IBOutlet private var savePanelAccessoryView: NSView?
+    @IBOutlet private weak var savePanelAccessoryView: NSView?
     
     private var readingEncoding: String.Encoding  // encoding to read document file
     private var needsShowUpdateAlertWithBecomeKey = false
@@ -509,10 +509,10 @@ final class Document: NSDocument, EncodingHolder {
                 permissions = ((try? FileManager.default.attributesOfItem(atPath: originalPath))?[.posixPermissions] as? UInt16) ?? 0  // FILE_READ
             }
             if permissions == 0 {
-                permissions = 0644  // ???: Is the default permission really always 644?
+                permissions = 0o644  // ???: Is the default permission really always 644?
             }
             permissions |= S_IXUSR
-            attributes[FileAttributeKey.posixPermissions.rawValue] = permissions
+            attributes[FileAttributeKey.posixPermissions.rawValue] = NSNumber(value: permissions)
         }
         
         return attributes