+++ CHANGELOG.md
@@ -8,6 +8,7 @@ develop
 ### Fixes
 
 - [beta] Fix an issue where application crashed after user turned the “Give execute permission” checkbox in the save panel on.
+- [beta] Fix an issue where text layout orientation was not restored from the last session.
 - [beta] Improve general stability.