+++ CHANGELOG.md
@@ -15,6 +15,7 @@ develop
 - [beta] Fix an issue where application crashed after user turned the “Give execute permission” checkbox in the save panel on.
 - [beta] Fix an issue where toggling status bar visibility didn't work.
 - [beta] Fix an issue where text layout orientation was not restored from the last session.
+- [beta] Fix an issue where key binding modification in pereferences could fail.
 - [beta] Fix text wrapping behavior with vertical orientation.
 - [beta] Improve general stability.