+++ CHANGELOG.md
@@ -14,6 +14,7 @@ develop
 
 - Fix an issue where page guide remained after toggling page guide visibility.
 - [beta] Fix an issue where application crashed after user turned the “Give execute permission” checkbox in the save panel on.
+- [beta] Fix an issue where application crashed when file opens via Service.
 - [beta] Fix an issue where toggling status bar visibility didn't work.
 - [beta] Fix an issue where text layout orientation was not restored from the last session.
 - [beta] Fix an issue where key binding modification in pereferences could fail.