+++ CHANGELOG.md
@@ -19,6 +19,7 @@ develop
 - [beta] Fix an issue where text layout orientation was not restored from the last session.
 - [beta] Fix an issue where key binding modification in pereferences could fail.
 - [beta] Fix an issue where submenus in the Script menu displayed in the menu key bindings setting view.
+- [beta] Fix an issue where “Comment always from line head” option didn't reflect user state.
 - [beta] Fix text wrapping behavior with vertical orientation.
 - [beta] Improve general stability.